requirements.txt
# Requirements for s3-aurora-bq-airflow project
# For Docker and cloud deployment

# Core data processing
meltano==3.4.0
pandas>=2.0.0
numpy>=1.24.0

# Database connections
pymysql>=1.0.0
sqlalchemy>=2.0.0

# AWS services
boto3>=1.26.0
botocore>=1.29.0

# Google Cloud services
google-cloud-bigquery>=3.11.0
google-auth>=2.20.0

# Meltano extractors and loaders
pipelinewise-tap-mysql>=1.5.0
target-bigquery>=0.12.0

# File handling
fsspec>=2023.6.0
s3fs>=2023.6.0

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.0
jsonschema==4.23

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
